refactor(domain): clean archtecture folder structure
