+++ src/domain/carrier/application/repositories/orders-repository.ts
@@ -1,4 +1,4 @@
-import { Order } from '@/domain/entities/order'
+import { Order } from '@/domain/carrier/enterprise/entities/order'
 
 export interface OrdersRepository {
   create(order: Order): Promise<void>
