+++ src/domain/carrier/application/use-cases/create-order.ts
@@ -1,5 +1,5 @@
-import { Order } from '@/domain/entities/order'
-import { OrdersRepository } from '@/domain/repositories/orders-repository'
+import { Order } from '@/domain/carrier/enterprise/entities/order'
+import { OrdersRepository } from '@/domain/carrier/application/repositories/orders-repository'
 import { UniqueEntityID } from '@/core/entities/unique-entity-id'
 
 interface CreateOrderUseCaseRequest {
