+++ src/domain/carrier/application/use-cases/update-order-delivery-status.ts
@@ -1,4 +1,4 @@
-import { Order, Status } from '@/domain/entities/order'
+import { Order, Status } from '@/domain/carrier/enterprise/entities/order'
 
 interface UpdateOrderDeliveryStatusRequestUseCase {
   order: Order
