+++ src/domain/carrier/enterprise/entities/manager.ts
