+++ vitest.config.ts
@@ -1,4 +1,3 @@
-// vitest.config.ts
 import { defineConfig } from 'vitest/config'
 import tsconfigPaths from 'vite-tsconfig-paths'
